+++ .local/bin/wallheaven.py
@@ -15,9 +15,9 @@ def init():
     category_code = get_category_code()
     purity_code = get_purity_code()
     global path
-    path = "/home/john/.scripts/image"
+    path = "/home/john/Pictures"
     global BASE_URL
-    BASE_URL = f"https://wallhaven.cc/api/v1/search?apikey={API_KEY}"+f"&q={keyword}"+f"&purity={category_code}"+f"&purity={purity_code}"+f"&atleast={res}"+f"&ratios={ratio}"+f"&page=1"
+    BASE_URL = f"https://wallhaven.cc/api/v1/search?apikey={API_KEY}"+f"&q={keyword}"+f"&purity={category_code}"+f"&purity={purity_code}"+f"&atleast={res}"+f"&ratios={ratio}"+f"&sorting=random"+f"&page=1"
 def search():
     query = input("Put the keyword you want search: ")
     return query
@@ -28,7 +28,6 @@ def resolution():
     while res not in ('720p', '1080p', '2160p'):
         print("Wrong resolution input")
         res = input("enter resolution: ").lower()
-        print("\n")
 
     resolution_tags = {'720p': '1280x720', '1080p': '1920x1080', '2160p': '3840x2160'}
     resolution_code = resolution_tags[res]
@@ -39,7 +38,6 @@ def ratios():
     while ratio not in ('16:9', '21:9', 'all'):
         print('Wrong ratio input')
         ratio = input("Enter ratio: ").lower()
-        print('\n')
     ratio_tags = {'16:9': '16x9', '21:9': '21x9', 'all': 'landscape'}
     ratio_code = ratio_tags[ratio]
     return ratio_code
@@ -54,7 +52,7 @@ def get_category_code():
 def get_purity_code():
     purity_tags = {'sfw': '100', 'sketchy': '010', 'nsfw': '001', 'ws': '110', 'wn': '101', 'sn': '011', 'all': '111'}
 
-    purity_code = purity_tags['all']
+    purity_code = purity_tags['ws']
 
     return purity_code
 
@@ -68,9 +66,12 @@ def image_download(url):
     try:
         if r.status_code == 200:
             content = r.content
-            with open(savefile, 'wb') as f:
-                f.write(content)
-            print("Downloading: "+ filename)
+            if not os.path.isfile(savefile):
+                with open(savefile, 'wb') as f:
+                    f.write(content)
+                print("Downloading: "+ filename)
+            else:
+                print(filename, 'is already exists skipping download')
         else:
             print("Failed to download ",filename , r.status_code)
             return None
@@ -93,16 +94,8 @@ def main():
         BASE_URL = BASE_URL[:-1]+f"{page}"
         print(BASE_URL, "is crawing")
         wall = requests.get(BASE_URL, headers=headers).json()["data"]
-
-   #     for i in range(len(wall)):
-   #         img = wall[i]["path"]
-   #         image.append(img)
-
-   # for imgs in image:
-   #     image_download(imgs)
         
-        
-        with ThreadPoolExecutor(50) as t:
+        with ThreadPoolExecutor(max_workers=5) as t:
             for i in range(len(wall)):
                 img = wall[i]["path"]
                 t.submit(image_download, url=img)
